Separate actor behaviour and data
