+++ src/actor_prototypes.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "typedef.h"
 #include "collision.h"
-#include "actor_types.h"
+#include "actor_data.h"
 #include "asset_types.h"
 
 struct ActorPrototype {
