--- src/actor_types.cpp
@@ -1,2 +1,0 @@
-#include "actor_types.h"
-
