+++ src/player.h
@@ -1,17 +1,7 @@
 #pragma once
-#include "typedef.h"
-#include "asset_types.h"
-#include "actor_reflection.h"
+#include "player_data.h"
 #define GLM_FORCE_RADIANS
 #include <glm.hpp>
-
-#pragma region Player
-enum PlayerType : TActorSubtype {
-	PLAYER_TYPE_SIDESCROLLER,
-	PLAYER_TYPE_OVERWORLD,
-
-	PLAYER_TYPE_COUNT
-};
 
 enum PlayerWeaponType : u8 {
 	PLAYER_WEAPON_BOW,
@@ -29,48 +19,6 @@ enum PlayerModeBits : u8 {
 	PLAYER_MODE_DODGE,
 };
 
-struct PlayerData {
-	SoundHandle jumpSound;
-	SoundHandle damageSound;
-	SoundHandle gunSound; // TEMP!
-};
-
-ACTOR_SUBTYPE_PROPERTIES(PlayerData,
-	ACTOR_SUBTYPE_PROPERTY_ASSET(PlayerData, jumpSound, ASSET_TYPE_SOUND, 1),
-	ACTOR_SUBTYPE_PROPERTY_ASSET(PlayerData, damageSound, ASSET_TYPE_SOUND, 1),
-	ACTOR_SUBTYPE_PROPERTY_ASSET(PlayerData, gunSound, ASSET_TYPE_SOUND, 1)
-);
-
-struct PlayerFlags {
-	u8 aimMode : 2;
-	bool slowFall : 1;
-	bool doubleJumped : 1;
-	bool airDodged : 1;
-	u8 mode : 4;
-};
-
-struct PlayerState {
-	PlayerFlags flags;
-
-	u16 wingCounter;
-	u16 wingFrame;
-	u16 shootCounter;
-	
-	u16 modeTransitionCounter;
-	u16 staminaRecoveryCounter;
-};
-
-struct PlayerOverworldData {
-
-};
-
-ACTOR_SUBTYPE_PROPERTIES(PlayerOverworldData)
-
-struct PlayerOverworldState {
-	glm::ivec2 facingDir;
-	u16 movementCounter;
-};
-
 struct Actor;
 struct Damage;
 
@@ -82,5 +30,3 @@ namespace Game {
 	bool PlayerInvulnerable(Actor* pPlayer);
 	void PlayerTakeDamage(Actor* pPlayer, const Damage& damage, const glm::vec2& enemyPos);
 }
-
-DECLARE_ACTOR_EDITOR_DATA(player)
