+++ src/spawner_data.h
@@ -1,0 +1,48 @@
+#pragma once
+#include "asset_types.h"
+#include "actor_reflection.h"
+
+enum SpawnerType : TActorSubtype {
+	SPAWNER_TYPE_EXP,
+	SPAWNER_TYPE_ENEMY,
+	SPAWNER_TYPE_LOOT,
+
+	SPAWNER_TYPE_COUNT
+};
+
+struct ExpSpawnerData {
+	ActorPrototypeHandle large;
+	ActorPrototypeHandle small;
+};
+
+ACTOR_SUBTYPE_PROPERTIES(ExpSpawnerData,
+	ACTOR_SUBTYPE_PROPERTY_ASSET(ExpSpawnerData, large, ASSET_TYPE_ACTOR_PROTOTYPE, 1),
+	ACTOR_SUBTYPE_PROPERTY_ASSET(ExpSpawnerData, small, ASSET_TYPE_ACTOR_PROTOTYPE, 1)
+);
+
+struct ExpSpawnerState {
+	u16 remainingValue;
+};
+
+struct EnemySpawnerData {
+};
+
+ACTOR_SUBTYPE_PROPERTIES(EnemySpawnerData);
+
+struct LootSpawnerData {
+	u8 typeCount;
+	u8 spawnRates[4];
+	ActorPrototypeHandle types[4];
+};
+
+ACTOR_SUBTYPE_PROPERTIES(LootSpawnerData,
+	ACTOR_SUBTYPE_PROPERTY_SCALAR(LootSpawnerData, typeCount, U8, 1),
+	ACTOR_SUBTYPE_PROPERTY_SCALAR(LootSpawnerData, spawnRates, U8, 4),
+	ACTOR_SUBTYPE_PROPERTY_ASSET(LootSpawnerData, types, ASSET_TYPE_ACTOR_PROTOTYPE, 4)
+);
+
+ACTOR_EDITOR_DATA(Spawner,
+	{ "exp_spawner", GET_SUBTYPE_PROPERTIES(ExpSpawnerData) },
+	{ "enemy_spawner", GET_SUBTYPE_PROPERTIES(EnemySpawnerData) },
+	{ "loot_spawner", GET_SUBTYPE_PROPERTIES(LootSpawnerData) }
+)
